refactor: add getDate fn
