+++ JS/views/components/card/CardView.js
@@ -1,3 +1,4 @@
+import { getTime } from "../../../controllers/utils/getDate.js";
 import { tagList } from "../../../data.js";
 import { BothCardView } from "./BothCardView.js";
 import { CardFooterView } from "./CardFooterView.js";
@@ -6,34 +7,8 @@ import { OnlyImgCardView } from "./OnlyImgCardView.js";
 import { OnlyTextCardView } from "./OnlyTextCardView.js";
 
 export const CardView = ({ nickname, content, imageUrl, like, createdTime, id, tagId }) => {
-  // * db상의 tag와 fe data의 index 차이 보정
-
   const profileImg = tagId ? tagList[tagId].img : tagList[0].img;
-  const originalDate = new Date(createdTime);
-
-  // 같은날 올린 글 - HH:mm
-  const sameDayOptions = {
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: false,
-  };
-
-  // 다른날 올린 글 - MM/dd HH:mm
-  const differentDayOptions = {
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: false,
-  };
-
-  const currentDate = new Date();
-  let formattedTime;
-  if (originalDate.getFullYear() === currentDate.getFullYear() && originalDate.getMonth() === currentDate.getMonth() && originalDate.getDate() === currentDate.getDate()) {
-    formattedTime = originalDate.toLocaleString("en-US", sameDayOptions);
-  } else {
-    formattedTime = originalDate.toLocaleString("en-US", differentDayOptions).replace(",", "");
-  }
+  const formattedTime = getTime(createdTime);
 
   return `
 <section class="wagle__main__card" id="${id}">
